expr: Fix RHS needs-paren detection
The issue was two fold.

First: It does not matter if parent/RHS expressions have the same
operator, (before) but it does matter if they have the same
precedent (now).

Second: Even with matching parent/RHS precedents, parens might not be
needed. The second condition is that the _parent_ is not associative.
Consider:

Correct  1 - (2 + 1)  ==  -2  ==  S (N 1) (A (N 2) (N 1))
  Wrong  1 - 2 + 1    ==   0  /=  S (N 1) (A (N 2) (N 1))

Note, interestingly, that this second case exclusively depends upon the
parent, meaning it does not matter if the child is associative or not.
Consider:

Correct  1 + 2 - 1    ==  2
Correct  1 + (2 - 1)  ==  2

The reason for this might be more clear when subtraction is
made into explicit addition of negative numbers:

DIFFERENT!
1-(2+1)  ->  1 +(-2) +(-1)
1-2+1    ->  1 +(-2) +1

SAME!
1+2-1    ->  1 +2 +(-1)
1+(2-1)  ->  1 +2 +(-1)