trie: fixes: Indexing overlaps. Prefixes that are words.
Indexing Overlaps: When indexing a word whose path already
fully exists in the trie then we must mark the node where the word
finishes as being a word.

Prefixes that are words: Totally new `findTrieWithPrefix`. This fixes
an error wherein search would fail to return words that matched the
given prefix. Instead of returning a subTrie we return a pruned trie
that includes the prefix path (but none of its branches). Furthermore
the prefix path nodes are marked as not words since for this prefix
tree they are not eligible (the prefix goes beyond them thus they
are not the words being looked for). NOTE Arguably this function is
heavily tied to `search` and not intuitive in its own right.
Not sure yet. At any rate the bug is fixed now.