quine: Solve.
Full disclosure: I was quite lost on how to solve this before looking at
the solution submitted by @balachandranc [1]. All I have done is
simplify it. I knew the solution would be "clever" meaning short but
with the right mindset. I became hung up on trying to avoid writing
the code-rewriting in the code-writing (that causes an infinite loop).

The solution is trivial: Keep the code-rewrite on the last line. Print
it above but first append the string to itself. Beautiful and simple.

[1] https://github.com/neongreen/haskell-ex/pull/30/files