json-print: Fix quote escape.
The issue was two-fold. First the direct problem was that
escaping a string is a matter of transmitting a literal slash prepended
to the quote, but I was supplying two.

The second issue was that I was susceptible to this problem because
my "testing" method was to take stringified values from my Haskell
demo and try to parse them in node. However I was using the value
from Haskell that I passed through `putStrLn` Which reduces the
slashes to appear correct. Ad-hoc testing methodology results in pain.