quickcheck: Expose all test cases.
`test` method used to only expose test cases that invalidated the
property being tested. However for certain modes (e.g. verbose) we
want to show all test cases, fail or pass. To achieve this we replace
Maybe with Either. Test cases that invalidate the property are in the
Left constructor, test cases that pass are in Right.